+++ testapp/src/main/AndroidManifest.xml
@@ -13,16 +13,16 @@
         android:theme="@style/AppTheme"
         tools:ignore="GoogleAppIndexingWarning">
 
-        <activity android:name="com.rakuten.tech.mobile.testapp.ui.miniapplist.MiniAppListActivity"/>
-
-        <activity android:name="com.rakuten.tech.mobile.testapp.ui.base.BaseActivity"/>
-        <activity android:name="com.rakuten.tech.mobile.testapp.ui.display.MiniAppDisplayActivity">
+        <activity android:name="com.rakuten.tech.mobile.testapp.ui.miniapplist.MiniAppListActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity android:name="com.rakuten.tech.mobile.testapp.ui.base.BaseActivity"/>
+        <activity android:name="com.rakuten.tech.mobile.testapp.ui.display.MiniAppDisplayActivity"/>
 
         <meta-data
             android:name="com.rakuten.tech.mobile.miniapp.BaseUrl"
