+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/base/BaseFragment.kt
@@ -1,0 +1,19 @@
+package com.rakuten.tech.mobile.testapp.ui.base
+
+import androidx.fragment.app.Fragment
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.SupervisorJob
+import kotlin.coroutines.CoroutineContext
+
+open class BaseFragment : Fragment(), CoroutineScope {
+    private val job: Job = SupervisorJob()
+    override val coroutineContext: CoroutineContext
+        get() = Dispatchers.IO + job
+
+    override fun onDestroy() {
+        super.onDestroy()
+        job.cancel()
+    }
+}
