+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/miniapplist/MiniAppListViewModel.kt
@@ -1,6 +1,5 @@
 package com.rakuten.tech.mobile.testapp.ui.miniapplist
 
-import android.content.Context
 import android.webkit.WebView
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
@@ -8,7 +7,6 @@ import androidx.lifecycle.ViewModel
 import com.rakuten.tech.mobile.miniapp.MiniApp
 import com.rakuten.tech.mobile.miniapp.MiniAppInfo
 import com.rakuten.tech.mobile.miniapp.MiniAppSdkException
-import com.rakuten.tech.mobile.miniapp.MiniAppView
 
 class MiniAppListViewModel constructor(
     private val miniapp: MiniApp
@@ -39,11 +37,6 @@ class MiniAppListViewModel constructor(
         } catch (error: MiniAppSdkException) {
             _errorData.postValue((error.message))
         }
-    }
-
-    suspend fun obtainMiniAppView(miniAppInfo: MiniAppInfo, context: Context) {
-        val miniAppView: MiniAppView = miniapp.create(miniAppInfo)
-        _miniAppView.postValue(miniAppView.obtainView(context))
     }
 
 }
